test(#1483): feature-gate read-sstable tests + extract shared fixture helpers (roborev)
